+++ helpers/exceptions.rb
@@ -1,4 +1,3 @@
 class InvalidScrollError < StandardError; end
 class UnknownScrollError < StandardError; end
-class InstallScrollError < StandardError; end
 class AlreadyInstalledError < StandardError; end
