+++ main.rb
@@ -32,6 +32,9 @@ EM.run do
   $scheduler = Rufus::Scheduler::EmScheduler.start_new
   Console.show 'Scheduler loaded !', 'success'
 
+  #SampleScroll = ScrollInstaller.new('Java')
+  #SampleScroll.install
+
   SampleScroll = ScrollInstaller.new('Bukkit', {folder: 'minecraft_test', user:'dernise', port:25568})
   SampleScroll.install
