--- modules/monitor/initializer.rb
@@ -1,5 +1,0 @@
-require_relative 'process'
-require_relative 'system'
-require_relative 'monitor'
-
-include Monitoring::System
