+++ modules/services/initializer.rb
@@ -1,0 +1,5 @@
+require_relative 'service'
+require_relative 'system'
+require_relative 'monitor'
+
+include Monitoring::System
