+++ scrolls/Java/java.rb
@@ -1,13 +1,14 @@
 class Java < Scroll
-  def initialize
+  def initialize(options = {})
     self.name = 'Java'
     self.author = 'Dernise'
     self.version = '1.7'
     self.homepage = 'http://wwww.edenservers.fr'
+    self.dependable = false
     super
   end
 
-  def install(options = {})
+  def install
     System.apt_get('update')
     System.apt_get('install', 'openjdk-7-jre')
     register('/usr/bin', 'java -version')
